+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 # 1.1.0
 
+- Bash completion of sub-command names:
+
+        $ complete -C 'jirash --bash-completion' jirash    # setup bash completion
+        $ jirash cre<TAB>    # completes to "jirash createissue"
+
 - '-a|--assignee' optiotn to 'jirash createissue'
 - `jirash filters`
 - `jirash user USERNAME`