+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 # 1.1.0
 
+- Monkey-patch xmlrpclib to not bork on invalid UTF-8 XML response from Jira's
+  XML-RPC API. Side-effect is that an invalid text field will have decode errors
+  replaced with '?'.
 - `jirash statuses`
 - `jirash issues TERMS...`, `jirash issues -f FILTER`
 - Bash completion of sub-command names: