+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 # 1.1.0
 
+- [issue #1] Fix create issue with no given assignee
 - `jirash resolve FOO-123` No support for setting the resolution (e.g. "Fixed"
   vs. "Won't Fix") because that's not possible via the Jira APIs.
 - `jirash resolutions`